% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{allocate_reservoir_to_river}
\alias{allocate_reservoir_to_river}
\title{Allocate each reservoir to nearest river reach within a given subbasin}
\usage{
allocate_reservoir_to_river(hybas_id, riv_i, res_i)
}
\arguments{
\item{hybas_id}{from \code{data(otto)}}

\item{riv_i}{a subset of river reaches from \code{data(riv)}}

\item{res_i}{a subset of reservoirs from \code{data(res_max)}}
}
\value{
reservoir_near_river the reservoir with the snapped reservoirs
}
\description{
Allocate each reservoir to nearest river reach within a given subbasin
}
