% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw_api_rain.R
\name{apiRain}
\alias{apiRain}
\title{Get rain from api data}
\usage{
apiRain(catch, api, date = as.Date("2018-03-15"))
}
\arguments{
\item{catch}{catch output of \code{contributing_basins_res} or \code{contributing_basins_shape}}

\item{api}{output of \code{request_api_gauges}}
}
\value{
\code{api_means} is a geospatial dataframe with the mean api precipitation for each subbasin of the input \code{catch}
}
\description{
This function interpolates rain data using idw (inverse distance weighted) interpolation
}
