% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident_contributing_basins_gauges.R
\name{contributing_basins_at_geom}
\alias{contributing_basins_at_geom}
\title{Identify contributing basins - sf}
\usage{
contributing_basins_at_geom(geom = res_max[res_max$id_jrc == 25283, ])
}
\arguments{
\item{geom}{a geometry from an sf object (WGS84, UTM zone=24 south) to identify its contributing basins, e.g. a reservoir from \code{data(res_max)}}
}
\value{
a geospatial dataframe of all contributin subbasins
}
\description{
This function identifies contributing basins of an sf geospatial dataframe
}
