% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident_contributing_basins_gauges.R
\name{contributing_basins_at_res}
\alias{contributing_basins_at_res}
\title{Identify contributing basins - reservoir}
\usage{
contributing_basins_at_res(ID = 25283)
}
\arguments{
\item{ID}{id of a reservoir from \code{res_max}}
}
\value{
a geospatial dataframe of all contributing subbasins
}
\description{
This function identifies contributing basins of a reservoir from \code{res_max}
}
