% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trmm_world.R
\name{get_trmm_world}
\alias{get_trmm_world}
\title{List and download trmm data in ftp}
\usage{
get_trmm_world(YEAR, MONTH, DAY)
}
\arguments{
\item{YEAR}{year}

\item{MONTH}{month}

\item{DAY}{and day to be downloaded}
}
\value{
files_world, contains the names of the available trmm files
}
\description{
List and download trmm data in ftp
}
