% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assim-data.R
\docType{data}
\name{hymet}
\alias{hymet}
\title{Time series of the stations described in columns \emph{name}, \emph{id} and \emph{geometry}. Values are given for \emph{evaporation}, \emph{precipitation} and \emph{runoff} time series in mm}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 683748 rows and 7 columns.}
\usage{
data(hymet)
}
\description{
Time series of the stations described in columns \emph{name}, \emph{id} and \emph{geometry}. Values are given for \emph{evaporation}, \emph{precipitation} and \emph{runoff} time series in mm
}
\keyword{datasets}
