% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identBasinsGauges.R
\name{identBasinsGauges}
\alias{identBasinsGauges}
\title{Contributing basins and rain gauges}
\usage{
identBasinsGauges(ID)
}
\description{
This function identifies contributing basins and surrounding rain gauges
}
