% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identBasinsGauges_shape.R
\name{identBasinsGauges_shape}
\alias{identBasinsGauges_shape}
\title{Contributing basins and rain gauges}
\usage{
identBasinsGauges_shape(shape, distGauges = 30)
}
\arguments{
\item{shape}{shapefile (WGS84, UTM zone=24) to identify its contributing basins and surrounding rain gauges, e.g. a reservoir from \code{data(res_max)}}

\item{distGauges}{distance in km around the contributing basins to look for rain gauges, defaults to 30}
}
\value{
a list with 6 elements:
\itemize{
\item \code{res} is the treated shape/reservoir,
\item \code{catch} is the catchment contributing to this shape/reservoir,
\item \code{catch_km2} gives the area of the catchment in square kilometers,
\item \code{catch_buffer} is a shapefile of a buffer zone of the chosen size around the catchment,
\item \code{gauges_catch} is a point shapefile with the rain gauges within "catch_buffer" and
\item \code{routing} is logical, indicating if routing can be done (TRUE when the shape/reservoir receives water from upstream subbasins)
}
}
\description{
This function identifies contributing basins and surrounding rain gauges
}
