% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwRain.R
\name{idwRain}
\alias{idwRain}
\title{Interpolate rain data}
\usage{
idwRain(catch, gauges_catch, api, distGauges = 30, ID)
}
\arguments{
\item{catch}{catch output of \code{contributing_basins_res} or \code{contributing_basins_shape}}

\item{gauges_catch}{output of \code{rain_gauges_catch}}

\item{api}{output of \code{request_api_gauges}}
}
\value{
a dataframe with the mean precipitation of the catchment and the reservoir of each requested day.
}
\description{
This function interpolates rain data using idw (inverse distance weighted) interpolation
}
