% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assim-data.R
\docType{data}
\name{p_gauges_saved}
\alias{p_gauges_saved}
\title{Meta data of precipitation gauges in Ceará}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 230 rows and 18 columns.}
\source{
http://api.funceme.br/help/doc/servicos-publicos-v1
}
\usage{
data(p_gauges_saved)
}
\description{
Meta data of precipitation gauges in Ceará
}
\keyword{datasets}
