% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_api_rain}
\alias{plot_api_rain}
\title{Plot idw interpolation}
\usage{
plot_api_rain(list_api_rain)
}
\arguments{
\item{list_api_rain}{output of api_rain_raster}
}
\description{
This function plots the interpolated precipitation in the contributing basins
}
