% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contributing_basins}
\alias{plot_contributing_basins}
\title{Plot contributing basins}
\usage{
plot_contributing_basins(catch, shape)
}
\arguments{
\item{catch}{output from \code{contributing_basins_shape} or \code{contributing_basins}}
}
\description{
This function plots the identified contributing basins of contributing_basins_shape and contributing_basins
}
