% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gauges_catch}
\alias{plot_gauges_catch}
\title{Plot rain gauges}
\usage{
plot_gauges_catch(catch, gauges_catch, distGauges = 30)
}
\arguments{
\item{catch}{output of \code{contributing_basins_res} or \code{contributing_basins_shape}}

\item{gauges_catch}{output of \code{rain_gauges_catch}}

\item{distGauges}{distance in km around the contributing basins to look for rain gauges as it was used in identBasinsGauges, defaults to 30}
}
\description{
This function plots the identified rain gauges of identBasinsGauges
}
