% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBasins}
\alias{plotBasins}
\title{Plot contributing basins}
\usage{
plotBasins(list_BG)
}
\arguments{
\item{list_BG}{output of identBasinsGauges}
}
\description{
This function plots the identified contributing basins of identBasinsGauges
}
