% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotIDW}
\alias{plotIDW}
\title{Plot idw interpolation}
\usage{
plotIDW(list_output, list_idw)
}
\arguments{
\item{list_output}{output of identBasinsGauges}

\item{list_idw}{output of idwRain}
}
\description{
This function plots the interpolated precipitation in the contributing basins
}
