% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTRMM}
\alias{plotTRMM}
\title{Plot trmm of subbasins}
\usage{
plotTRMM(trmm_means)
}
\arguments{
\item{trmm_means}{output of trmmRain, geospatial dataframe with mean TRMM precipitation of each field of a given sf geometric object}
}
\description{
This function plots the averaged precipitation in all subbasins of the contributing basins
}
