% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assim-data.R
\docType{data}
\name{postos}
\alias{postos}
\title{Meta data of stations for precipitation, evaporation and runoff in Jaguaribe basin, Ceará}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 18 rows and 6 columns.}
\usage{
data(postos)
}
\description{
Meta data of stations for precipitation, evaporation and runoff in Jaguaribe basin, Ceará
}
\keyword{datasets}
