% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident_contributing_basins_gauges.R
\name{rain_gauges_catch}
\alias{rain_gauges_catch}
\title{Identify Rain gauges of the catchment}
\usage{
rain_gauges_catch(catch, distGauges = 30)
}
\arguments{
\item{catch}{output of \code{contributing_basins_res} or \code{contributing_basins_shape}}

\item{distGauges}{distance in km around the contributing basins to look for rain gauges}
}
\value{
a point shapefile with the rain gauges within the chosen distance (distGauges)
}
\description{
This function identifies rain gauges within a certain distance around the catchment
}
