% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestGauges.R
\name{requestGauges}
\alias{requestGauges}
\title{Request rain data from FUNCEME api}
\usage{
requestGauges(requestDate, Ndays, list_output)
}
\description{
This function requests api rain data for selected rain gauges
}
