% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assim-data.R
\docType{data}
\name{res_max}
\alias{res_max}
\title{Maximum extent of reservoirs in Ceará}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 22960 rows and 12 columns.}
\source{
database
}
\usage{
data(res_max)
}
\description{
Maximum extent of reservoirs in Ceará
}
\keyword{datasets}
