% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_model2.R
\name{res_model2}
\alias{res_model2}
\title{Model reservoirs}
\usage{
res_model2(ID = 33443, start = as.Date("2000-01-16"),
  end = as.Date("2000-01-18"))
}
\arguments{
\item{ID}{ID of the reservoir for which the contributing basins shall be modelled}

\item{start}{date at which model run shall start, default: as.Date("2000-01-10")}

\item{end}{date at which model run shall end, default: as.Date("2000-01-15")}
}
\value{
table with vol_0 (volume at the beginning of the timestep t), q_in_m3 (inflow in m3), q_out_m3 (outflow in m3) and vol_1 (volume at the end of the timestep) for each reservoir and for each subbasin
}
\description{
This function models runoff concentration for the contributing basins of a shapefile
}
