% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_evaporation.R
\name{reservoir_evaporation}
\alias{reservoir_evaporation}
\title{Evaporation from reservoir surface}
\usage{
reservoir_evaporation()
}
\description{
Calculate evaporation from reservoir surface to include in \code{reservoir_model}
}
