% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_withdrawl.R
\name{reservoir_withdrawl}
\alias{reservoir_withdrawl}
\title{Water withdrawl from reservoirs}
\usage{
reservoir_withdrawl()
}
\description{
Calculate water withdrawl from the reservoirs, 1% of maximum volume per day, to include in \code{reservoir_model}
}
