% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_graph_2.R
\name{Routing}
\alias{Routing}
\title{Reservoir routing of strategic reservoirs ####}
\usage{
Routing()
}
\value{
the column \code{res_down} in the geospatial dataframe \code{res_max}
}
\description{
This function identifies which strategic reservoir drains into which stategic downstream reservoir
}
