% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_non_strategic.R
\name{Routing_non_strat}
\alias{Routing_non_strat}
\title{Reservoir routing of non-strategic reservoirs ####}
\usage{
Routing_non_strat()
}
\value{
completes the column \code{res_down} in the geospatial dataframe \code{res_max}
}
\description{
This function identifies which non-strategic reservoir drains into which stategic downstream reservoir
}
