% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_strategic.R
\name{Routing_strat}
\alias{Routing_strat}
\title{Reservoir routing of strategic reservoirs ####}
\usage{
Routing_strat()
}
\value{
the column \code{res_down} in the geospatial dataframe \code{res_max}
}
\description{
This function identifies which strategic reservoir drains into which stategic downstream reservoir
}
