% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_contributing_areas.R
\name{runoff_contributing}
\alias{runoff_contributing}
\title{Estimate runoff contributing areas of each reservoir ####}
\usage{
runoff_contributing(start = as.Date("1960-01-01"),
  end = as.Date("1989-12-31"))
}
\arguments{
\item{start}{default: as.Date("1960-01-01")}

\item{end}{default: as.Date("1990-01-01")}
}
\value{
columns runoff_est and runoff_adapt in dataframe res_max
}
\description{
This function estimates the area of directly contributing runoff for all reservoirs, based on the mean yearly runoff of 1960-1990
}
