% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_contributing_area.R
\name{runoff_contributing_area}
\alias{runoff_contributing_area}
\title{Estimate runoff contributing areas of each reservoir ####}
\usage{
runoff_contributing_area(start = as.Date("1960-01-01"),
  end = as.Date("1989-12-31"))
}
\arguments{
\item{start}{default: as.Date("1960-01-01")}

\item{end}{default: as.Date("1989-12-31")}
}
\value{
columns runoff_contr_est and runoff_contr_adapt in dataframe res_max, runoff_contr_adapt adapts the estimated (theoretic) runoff contributing area to the actual subbasin area
}
\description{
This function estimates the area of directly contributing runoff for all reservoirs, based on the mean yearly runoff of 1960-1990
}
