% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{snap2riv}
\alias{snap2riv}
\title{Snap each reservoir to the nearest river reach}
\usage{
snap2riv(riv, res)
}
\arguments{
\item{riv}{from \code{data(riv)}}

\item{res}{from \code{data(res)}}
}
\value{
res_snap the snapped reservoirs
}
\description{
Snap each reservoir to the nearest river reach
}
