% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assim-data.R
\docType{data}
\name{time_series}
\alias{time_series}
\title{Time series of the stations described in \code{postos}, each station is a list object with its code a name}
\format{An object of class \code{list} of length 18.}
\usage{
data(time_series)
}
\description{
Time series of the stations described in \code{postos}, each station is a list object with its code a name
}
\keyword{datasets}
