% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identBasinsGauges.R
\name{identBasinsGauges}
\alias{identBasinsGauges}
\title{Contributing basins and rain gauges}
\usage{
identBasinsGauges(ID, distGauges = 30)
}
\arguments{
\item{ID}{id (due to JRC) of the reservoir of interest}

\item{distGauges}{distance in km around the contributing basins to look for rain gauges, defaults to 30}
}
\value{
list_BG is a list of
}
\description{
This function identifies contributing basins and surrounding rain gauges
}
