% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwRain.R
\name{idwRain}
\alias{idwRain}
\title{Interpolate rain data}
\usage{
idwRain(list_output, api)
}
\arguments{
\item{list_output}{output of identBasinsGauges}

\item{api}{output of requestGauges}
}
\description{
This function interpolates rain data using idw (inverse distance weighted) interpolation
}
