% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwRain.R
\name{idwRain}
\alias{idwRain}
\title{Interpolate rain data}
\usage{
idwRain(list_BG, api)
}
\arguments{
\item{list_BG}{output of identBasinsGauges}

\item{api}{output of requestGauges}
}
\value{
a list with 2 elements: "idwRaster" contains a raster of the interpolated precipitation for each requested day, "dailyRain_table" is a dataframe with the mean precipitation on the catchment and the reservoir of each requested day.
}
\description{
This function interpolates rain data using idw (inverse distance weighted) interpolation
}
