% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStratRes}
\alias{plotStratRes}
\title{Plot strategic reservoirs}
\usage{
plotStratRes(list_output)
}
\arguments{
\item{list_output}{output of identBasinsGauges}
}
\description{
This function plots the strategic reservoirs identified by resRouting
}
