% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTRMM}
\alias{plotTRMM}
\title{Plot trmm of subbasins}
\usage{
plotTRMM(list_trmm)
}
\arguments{
\item{list_trmm}{output of trmmRain}
}
\description{
This function plots the averaged precipitation in all subbasins of the contributing basins
}
