% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestGauges.R
\name{requestGauges}
\alias{requestGauges}
\title{Request rain data from FUNCEME api}
\usage{
requestGauges(requestDate, Ndays, list_BG)
}
\arguments{
\item{requestDate}{latest date of interest, e.g. as.Date("2018-03-15") or today()}

\item{Ndays}{number of days to go back in time from requestDate}

\item{list_BG}{output of identBasinsGauges}
}
\value{
a dataframe with the precipitation available for the requested dates and gauges
}
\description{
This function requests api rain data for selected rain gauges
}
