% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resRouting.R
\name{resRouting}
\alias{resRouting}
\title{Create reservoir routing scheme}
\usage{
resRouting(list_BG)
}
\arguments{
\item{list_BG}{the output list of identBasinsGauges}
}
\description{
This function creates a routing scheme for the strategic reservoirs, those on the main river course. For each reservoir the next reservoir downstream is identified.
}
