% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trmm_world.R
\name{trmmRain}
\alias{trmmRain}
\title{Get rain from trmm data}
\usage{
trmmRain(shape, files_world)
}
\arguments{
\item{shape}{sf geometric object in WGS 84, longlat}

\item{files_world}{output of get_trmm_world}
}
\value{
\code{trmm_means} is a geospatial dataframe with the mean TRMM precipitation for each field of the input shape
}
\description{
obtain precipitation on the reservoir and the mean of its catchment and each subbasin
}
